Decrease indentation with shift-tab

+++ epaint/src/text/font.rs
@@ -127,7 +127,7 @@ impl FontImpl {
 
             if c == '\t' {
                 if let Some(space) = self.glyph_info(' ') {
-                    glyph_info.advance_width = crate::text::TAB_SIZE * space.advance_width;
+                    glyph_info.advance_width = crate::text::TAB_SIZE as f32 * space.advance_width;
                 }
             }
